Move ess.diffraction.* into ess.powder
