+++ docs/api-reference/index.md
@@ -1,8 +1,10 @@
 # API Reference
 
-## Module Attributes
+## Esspowder
+
+### Module Attributes
 ```{eval-rst}
-.. currentmodule:: ess.diffraction
+.. currentmodule:: ess.powder
 
 .. autosummary::
    :toctree: ../generated/attributes
@@ -10,26 +12,7 @@
    providers
 ```
 
-## Top-level functions
-
-```{eval-rst}
-.. autosummary::
-   :toctree: ../generated/functions
-   :recursive:
-
-   crop_tof
-   filter_events
-   finalize_histogram
-   group_by_two_theta
-   lowpass
-   merge_all_pixels
-   normalize_by_monitor
-   normalize_by_proton_charge
-   normalize_by_vanadium
-   remove_bad_pulses
-```
-
-## Submodules
+### Submodules
 
 ```{eval-rst}
 .. autosummary::
@@ -37,8 +20,13 @@
    :template: module-template.rst
    :recursive:
 
-   external.powgen
-   powder
+   conversion
+   correction
+   filtering
+   grouping
+   smoothing
    types
    uncertainty
+
+   external.powgen
 ```
