+++ src/ess/powder/grouping.py
@@ -1,5 +1,6 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2023 Scipp contributors (https://github.com/scipp)
+"""Grouping and merging of pixels / voxels."""
 
 from scippneutron.conversion.graph import beamline
